Model_api/requirements.txt
fastapi==0.118.2
uvicorn==0.37.0
pymongo==3.12.0
python-dotenv==1.1.1
tensorflow==2.20.0
pillow==11.3.0
numpy==2.3.3
scikit-learn==1.7.2
python-multipart==0.0.20
